edit log

+++ app/src/main/java/com/szubov/healthmonitor/PatientVitals.java
@@ -2,9 +2,9 @@ package com.szubov.healthmonitor;
 
 public class PatientVitals {
     private float weight;
-    private int steps;
+    private short steps;
 
-    public PatientVitals(float weight, int steps) {
+    public PatientVitals(float weight, short steps) {
         this.weight = weight;
         this.steps = steps;
     }
